+++ cachix/src/Cachix/Client.hs
@@ -14,8 +14,9 @@ main = do
   env <- mkEnv cachixoptions
   case command of
     AuthToken token -> Commands.authtoken env token
-    Create name -> Commands.create env name
     GenerateKeypair name -> Commands.generateKeypair env name
     Push pushArgs -> Commands.push env pushArgs
+    WatchStore watchArgs name -> Commands.watchStore env watchArgs name
+    WatchExec pushArgs name cmd args -> Commands.watchExec env pushArgs name cmd args
     Use name useOptions -> Commands.use env name useOptions
     Version -> putText cachixVersion