+++ nix/stack-shell.nix
@@ -1,9 +1,7 @@
-{ ghc }:
 let 
   sources = (import ./sources.nix);
   pkgs = import sources.nixpkgs {} ;
 in pkgs.haskell.lib.buildStackProject {
-  inherit ghc;
   name = "cachix-stack-shell";
   buildInputs = [
     pkgs.lzma