+++ cachix/CHANGELOG.md
@@ -7,6 +7,47 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [0.6.0] - 2021-01-08
+
+### Changed
+
+- Watching nix store doesn't push .drv files anymore
+- `cachix push -w` has been moved to `cachix watch-store`
+- `cachix create` has been removed
+- Retries now take a multiple of seconds instead of multiple of 100ms
+
+### Added
+
+- watch-exec allows to run a command and push all new store paths added meanwhile
+- GHC 8.10 support
+
+### Fixed
+
+- Watching /nix/store now uses queue to bulk query what is missing in binarycache and
+  a queue for pushing
+- Instructions for NixOS trusted users were inaccurate
+- Retry fetching binary cache
+
+## [0.5.1] - 2020-11-09
+
+### Fixed
+
+- Regression: use auth token when using signing key with private caches 
+- Configure netrc even if cachix config doesn't exist
+
+## [0.5.0] - 2020-11-06
+
+### Added
+
+- Allow specifying output directory to write nix.conf and netrc files. 
+- Allow pushing without a Signing key using only auth token
+- Allow setting auth token via `$CACHIX_AUTH_TOKEN` shell variable
+
+### Fixed
+
+- Watch store command now pushes the full closure of each store path
+- Support groups when parsing trusted-users from nix.conf
+
 ## [0.3.8] - 2020-06-03
 
 ### Added